+++ R/lc_get_params.R
@@ -16,21 +16,37 @@
 #' @export
 #'
 #' @examples
-#' params <- lc_get_params(param='name')
+#' params <- lc_get_params(param='variable_info')
+#' params <- lc_get_params(param='metric_names')
 #' params <- lc_get_params(param='areaOfInterest')
 #' params <- lc_get_params(param='state')
 #' params <- lc_get_params(param='county')
 
 lc_get_params <- function(param = NULL) {
+  UUID <- DATE_DOWNLOADED <- METADATA <- FINAL_TABLE<- NULL
+  INDICATOR_CATEGORY <- METRIC_NAME <- AOI <- YEAR <- NULL
+  WEBTOOL_NAME <- METRIC_UNITS <- METRIC_DESCRIPTION <- DSID <- NULL
+  SOURCE_NAME <- SOURCE_URL <- UUID <- DATE_DOWNLOADED <- NULL
   resp <- jsonlite::fromJSON("https://api.epa.gov/StreamCat/lakes/metrics")$items
   if (param=='areaOfInterest'){
     params <- strsplit(stringr::str_sub(resp$aoi_param_info[[1]]$options,2,-2),",")[[1]]
     params <- c(gsub(" ","", params),'other')
-    params <- params[1:2]
     params <- params[order(params)]
-  }  else if(param == 'name') {
+    params <- params[!params %in% c('catrp100','wsrp100','other')]
+  }  else if(param == 'metric_names') {
     params <- resp$name_options[[1]][[1]]
     params <- params[order(params)]
+  } else if(param == 'variable_info') {
+    params <- httr2::request('https://api.epa.gov/StreamCat/streams/variable_info') |>
+      httr2::req_perform() |>
+      httr2::resp_body_string() |>
+      readr::read_csv() |> 
+      dplyr::select(-UUID,-DATE_DOWNLOADED,-METADATA) |> 
+      dplyr::rename(dataset=FINAL_TABLE,category=INDICATOR_CATEGORY, 
+                    metric=METRIC_NAME,aoi=AOI, year=YEAR, 
+                    short_description=WEBTOOL_NAME,units=METRIC_UNITS,
+                    long_description=METRIC_DESCRIPTION, dsid=DSID,
+                    source_name=SOURCE_NAME, source_URL=SOURCE_URL)
   } else if(param == 'region'){
     params <- resp$region_options[[1]][[1]]
     params <- params[order(params)]